+++ src/is/index.js
@@ -1,0 +1,5 @@
+import types from './types';
+
+export default {
+    ...types,
+};
