+++ src/is/types.js
